+++ package/src/components/AudioPlayer/Context/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from "./useNonNullableContext";
 export * from "./usePlaybackContext";
 export * from "./useTimeContext";
 export * from "./useTrackContext";
