+++ package/src/components/AudioPlayer/Context/hooks/useAudioAttrsContext.ts
@@ -2,7 +2,7 @@ import {
   AudioAttrsContext,
   audioAttrsContext,
 } from "@/components/AudioPlayer/Context/AudioAttrsContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const useAudioAttrsContext = (): AudioAttrsContext =>
   useNonNullableContext(audioAttrsContext);
