+++ package/src/components/AudioPlayer/Context/hooks/usePlaybackContext.ts
@@ -2,7 +2,7 @@ import {
   PlaybackContext,
   playbackContext,
 } from "@/components/AudioPlayer/Context/PlaybackContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const usePlaybackContext = (): PlaybackContext =>
   useNonNullableContext(playbackContext);
