+++ package/src/components/AudioPlayer/Context/hooks/useResourceContext.ts
@@ -2,7 +2,7 @@ import {
   ResourceContext,
   resourceContext,
 } from "@/components/AudioPlayer/Context/ResourceContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const useResourceContext = (): ResourceContext =>
   useNonNullableContext(resourceContext);
