+++ package/src/components/AudioPlayer/Context/hooks/useTimeContext.ts
@@ -2,7 +2,7 @@ import {
   TimeContext,
   timeContext,
 } from "@/components/AudioPlayer/Context/TimeContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const useTimeContext = (): TimeContext =>
   useNonNullableContext(timeContext);
