+++ package/src/components/AudioPlayer/Context/hooks/useTrackContext.ts
@@ -2,7 +2,7 @@ import {
   TrackContext,
   trackContext,
 } from "@/components/AudioPlayer/Context/TrackContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const useTrackContext = (): TrackContext =>
   useNonNullableContext(trackContext);
