+++ package/src/components/AudioPlayer/Context/hooks/useUIContext.ts
@@ -1,4 +1,4 @@
 import { uiContext } from "@/components/AudioPlayer/Context/UIContext";
-import { useNonNullableContext } from "./useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 
 export const useUIContext = () => useNonNullableContext(uiContext);
