+++ package/src/components/AudioPlayer/Context/index.ts
@@ -8,3 +8,4 @@ export * from "./UIContext";
 export * from "./ResourceContext";
 export * from "./TimeContext";
 export * from "./providerProps";
+export * from "./hooks";
