+++ package/src/components/AudioPlayer/Interface/Controller/Drawer/SortablePlayList/Content/usePlayList.ts
@@ -4,9 +4,9 @@ import {
 } from "@/components/AudioPlayer/Context";
 import { CssTransitionProps } from "@/ui/CssTransition";
 import { UseSortableListItemProps } from "@/components/SortableList/useSortableListItem";
-import { useNonNullableContext } from "@/hooks/context/useNonNullableContext";
-import { useTrackContext } from "@/hooks/context/useTrackContext";
-import { useUIContext } from "@/hooks/context/useUIContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
+import { useTrackContext } from "@/components/AudioPlayer/Context/hooks/useTrackContext";
+import { useUIContext } from "@/components/AudioPlayer/Context/hooks/useUIContext";
 import { useCallback, useState } from "react";
 
 interface UsePlayListReturn {
