+++ package/src/components/AudioPlayer/Interface/Controller/Input/Progress/useProgressKeyDown.ts
@@ -1,5 +1,5 @@
-import { usePlaybackContext } from "@/hooks/context/usePlaybackContext";
-import { useResourceContext } from "@/hooks/context/useResourceContext";
+import { usePlaybackContext } from "@/components/AudioPlayer/Context/hooks/usePlaybackContext";
+import { useResourceContext } from "@/components/AudioPlayer/Context/hooks/useResourceContext";
 import React, { useCallback } from "react";
 
 export const useProgressKeyDown = (
