+++ package/src/components/Drawer/DrawerContent.tsx
@@ -1,4 +1,4 @@
-import { useNonNullableContext } from "@/hooks/context/useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 import {
   FC,
   KeyboardEvent,
