+++ package/src/components/Dropdown/DropdownTrigger.tsx
@@ -1,5 +1,5 @@
 import { ButtonHTMLAttributes, forwardRef } from "react";
-import { useNonNullableContext } from "@/hooks/context/useNonNullableContext";
+import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 import { dropdownContext } from "./DropdownContext";
 import { StyledBtn } from "@/ui/StyledBtn";
